Initialized a bunch of variables to shut up static analysis. Changed some functions to use their proper float, not doulble versions. Buffer tweaking for text.

+++ p_view.c
@@ -2424,7 +2424,7 @@ void ClientEndServerFrame(edict_t* ent)
 		bobtime *= 4;
 
 	bobcycle = (int)bobtime;
-	bobfracsin = fabs(sin(bobtime * M_PI));
+	bobfracsin = fabsf(sinf(bobtime * M_PI));  /* MetalGOd - it's all being truncated to a float, from a double so just use the float versions of fabs, and sin! */
 
 	// detect hitting the floor
 	P_FallingDamage(ent);
